+++ nSide/sfc/alt/ppu-performance/background/mode7.cpp
@@ -14,7 +14,7 @@ auto PPU::Background::render_mode7() -> void {
   int hofs = sclip<13>(self.regs.mode7_hoffset);
   int vofs = sclip<13>(self.regs.mode7_voffset);
 
-  int y = (self.regs.mode7_vflip == false ? self.vcounter() : 255 - self.vcounter());
+  int y = (self.regs.mode7_vflip == false ? (int)self.vcounter() : 255 - self.vcounter());
 
   uint16_t* mosaic_x;
   uint16_t* mosaic_y;
